#626 Split to separate components
