+++ next/src/components/layout/Sections.tsx
@@ -26,7 +26,7 @@ import ServicesSection from '@/src/components/sections/ServicesSection'
 import SortingGuideAccordionsSection from '@/src/components/sections/SortingGuideAccordionsSection'
 import SortingGuideSection from '@/src/components/sections/SortingGuideSection'
 import VacanciesSection from '@/src/components/sections/VacanciesSection'
-import WasteCollectionDays from '@/src/components/sections/WasteCollectionDays'
+import WasteCollectionDays from '@/src/components/sections/WasteCollectionDays/WasteCollectionDays'
 import WasteCollectionPointsSection from '@/src/components/sections/WasteCollectionPointsSection'
 import WasteRemovalCardsSection from '@/src/components/sections/WasteRemovalCardsSection'
 import WasteSortingCardsSection from '@/src/components/sections/WasteSortingCardsSection'
